+++ examples/tsne/responsive_tsne/responsive_tsne.h
@@ -58,8 +58,8 @@ static inline double sign(double x) { return (x == .0 ? .0 : (x < .0 ? -1.0 : 1.
 class ResponsiveTSNE
 {
 public:
-    void run(double* X, int N, int D, double* Y, int no_dims, double perplexity, double theta, int rand_seed,
-             bool skip_random_init, int max_iter, int stop_lying_iter, int mom_switch_iter, Config& config);
+    void run(double* X, size_t N, size_t D, double* Y, size_t no_dims, double perplexity, double theta, int rand_seed,
+             bool skip_random_init, size_t max_iter, size_t stop_lying_iter, size_t mom_switch_iter, Config& config);
     void updateSimilarity(Table *table, 
       vector<map<size_t, double>>& neighbors,
       vector<map<size_t, double>>& similarties,
